2003-07-23  Ziemowit Laski  <zlaski@apple.com>
This is the first check-in to the objc-improvements-branch (after its creation).
It is quite messy because it consists of two things:
  (1) Apple's extensive ObjC changes being merged back
  (2) A merge from TOT (objc-improvements-merge-20030722 tag).  This was done to
      verify that there are no regressions between objc-improvements-merge-20030722
      and objc-improvements-branch

Obviously, much more detailed ChangeLog entries (and diffs!) will need to be prepared when we merge
objc-improvements-branch into the mainline.  :-)  Also, most of this work will still have to be ANSI-fied
(now that the fatwah on ANSI syntax has been lifted.)

Many new test cases have been added into gcc/testsuite/objc.dg.  They all fall into one of
3 categories:
  (1) Test cases that succeed with the GNU as well as the NeXT runtime.
  (2) Test cases marked with 'target *-*-darwin*', since they rely on NeXT runtime features
  (3) Test cases which _should_ work on GNU, but for some reason do not.  (NB: They do work on Darwin.)
      This must be investigated further.

At this point, it would be very helpful if GNUStep developers could check out the objc-improvements-branch,
build the compiler, and identify any problems that arise.  If you come up with any fixes that affect only the GNU
mode (and leave the NeXT mode alone), please go ahead and check those in; I trust you. :-)  If a proposed fix
affects the NeXT mode in any way, I kindly ask that you post a diff to gcc-patches so that I may have a quick
look-see.

Thank you,

--Zem


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/objc-improvements-branch@69717 138bc75d-0d04-0410-961f-82ee72b054a4