Initial rewrite of <#pragma simd> clause parsing and handling.
The goal here is to use the existing OMP_CLAUSE* infrastructure by
adding OMP_CLAUSES in ``enum omp_clause_code'' and everything
associated with it.

We also want to reuse the OpenMP parsing and handling bits for the
private, firstprivate, lastprivate, and reduction clauses, since
according to the specs, the syntax and semantics are exactly the same.

This is still not working (at all), as we still need to engineer a
mechanism to move these clauses from gimple to ``struct loop'', at
least until Richi makes loop preservation happens much earlier.  As a
proof of concept, I'll probably hack something to move the OMP_CLAUSE*
information to the BBs and then to ``struct loop'' as is currently
implemented, at least to make sure things are still working.