Set register move costs so that they are never zero.
Only a singleton class can validly report that it costs nothing to
moved between registers of that class.

It must be possible to see the cost of moving from one register in a
class to another register in the same class. The cost of allocating a
register to satisfy constraints involving a sub and super class is
handled specially in IRA/LRA to take into account the zero cost of
allocating to a sub class and satisfying a constraint for a super
class (in constraint) and similarly for out constraint.