LRA and re-using argument save area
Attempt to educate LRA about stack slots created for arguments but not
filled in. This change allows argument registers to be pushed to their
natural slot as part of reloading which increases the chanced of folding
the initial push into the prologue save instruction for mips16. It also
reduces stack space in all cases.