clean up SHLIB so subshells are not needed
This changes the handling of SHLIB so that it is inlined into
DRIVER_DEFINES.

I think this is ok because SHLIB is defined in a Makefile fragment
that is included by the generated Makefile.

The rationale for this approach is that it simplifies some .o targets,
so that we can share more code.