Educate the loop cost calculations about mips16 register classes
The loop cost calculations are all based on the GENERAL_REGS class
which is not appropriate for mips16. This is mips16's fault but it
is a lot of work to change mips16.

The number of registers reserved for temporary expressions was also
fixed, this is now configurable and is reduced for mips16.