Enable LRA for the mips backend.
This gets LRA working but includes some workarounds.
A number of patterns need to accept the stack and frame pointer as well
so base registers now have to include these registers.

* Remove the load alternative from and<mode>3_mips16.
It is difficult to deal with the memory alternative when LRA is used. The
impact of removing it appears minimal but can be reviewed

* spill_class
Allowing the spill class logic to use any available register (as opposed
to just the call-clobbered non-mips16 registers) allows LRA to use normally
accesible registers should they happen to be free.  IRA can miss some
opportunities so this significantly improves reload code as reloads can be
effectively removed

Only disable the T registers for mips16 when using classic reload