Fix how --enable-host-shared adds -fPIC to compilation flags.
The --enable-host-shared configuration option is meant to add -fPIC to
compilation flags for code meant to run on the host.

Previously, I implemented this by detecting the --enable-host-shared
in each child "configure" script (as generated from configure.ac) and
conditionally appending -fPIC to makefile variables such as CFLAGS,
CXXFLAGS and LDFLAGS.

This worked when invoking "make" from each subdirectory, but not when
running "make" in the top-level directory: the top-level make would
pass in "CFLAGS=-g -O2" to the child "makes", overriding their CFLAGS,
thus erroneously omitting the "-fPIC" flag.

This patch instead adds a PICFLAG variable, set up in the child
configure scripts, to contain -fPIC if configured with
--enable-host-shared.  The PICFLAG variable is then added to an
appropriate variable in each child makefile to ensure that it is used
when building host code.

This ensures that -fPIC is present if requested, when building both
from the top-level directory, and from the child directories.

gcc/
	* Makefile.in: Set PICFLAG from configure script; add it to
	INTERNAL_CFLAGS.
	* configure.ac (--enable-host-shared): Set up PICFLAG rather
	than attempting to append -fPIC to CFLAGS, CXXFLAGS, LDFLAGS.
	* configure: Regenerate.

libcpp/
	* Makefile.in: Set PICFLAG from configure script; add it to
	ALL_CFLAGS.
	* configure.ac (--enable-host-shared): Set up PICFLAG rather
	than attempting to append -fPIC to CFLAGS, CXXFLAGS, LDFLAGS.
	* configure: Regenerate.

libdecnumber/
	* Makefile.in: Set PICFLAG from configure script; add it to
	ALL_CFLAGS.
	* configure.ac (--enable-host-shared): Set up PICFLAG rather
	than attempting to append -fPIC to CFLAGS, CXXFLAGS, LDFLAGS.
	* configure: Regenerate.

zlib/
	* Makefile.am: Add PICFLAG to libz_a_CFLAGS.
	* configure.ac (--enable-host-shared): Set up PICFLAG rather
	than attempting to append -fPIC to CFLAGS.
	* Makefile.in: Regenerate.
	* configure: Regenerate.