Use automatic dependencies.
Note that DEPFILES is built using CONFIG_LANGUAGES, and so this will
only work when building the jit in stages that expect it (and thus
have CONFIG_LANGUAGES containing "jit").

Note to self: I was using "make jit" in stage1's gcc subdirectory to
quickly build the code.  Given that this doesn't work with autodeps,
the better approach is to use --disable-bootstrap for doing this kind
of quick testing.

gcc/jit/
	* Make-lang.in: Rename JIT_OBJS to jit_OBJS.  Delete manual
	dependencies.