Implement source code location support (API change)
This patch implements gcc_jit_location, so that client code can mark
the various constructs with an underlying code location.  For example, a
JIT compiler for an interpreted scripting language may want to associate
the machine code with locations in the original script, allowing the
user to step through the JIT-compiled code in a debugger.

There is an API change in this patch:
gcc_jit_function_place_forward_label has gained a location as its
second parameter, and now has this signature:

 extern void
 gcc_jit_function_place_forward_label (gcc_jit_function *func,
				       gcc_jit_location *loc,
				       gcc_jit_label *lab);

An example of using this can be seen in jit.dg/test-fibonacci.c, where
the JIT code is set up to point the debugger at a comment within
that file containing C-like code.  It is possible to step through this
comment in the debugger:

  (gdb) break my_fibonacci
  (gdb) run
  Breakpoint 1, my_fibonacci (x=10) at testsuite/jit.dg/test-fibonacci.c:19
  19	FIRST_LINE + 3:   if (x < 2)
  (gdb) list
  14	0000000001111111111222222222233333333334444444444555555555566666666667
  15	1234567890123456789012345678901234567890123456789012345678901234567890
  16	FIRST_LINE + 0: int
  17	FIRST_LINE + 1: my_fibonacci (int x)
  18	FIRST_LINE + 2: {
  19	FIRST_LINE + 3:   if (x < 2)
  20	FIRST_LINE + 4:     return x;
  21	FIRST_LINE + 5:   else
  22	FIRST_LINE + 6:     return my_fibonacci (x - 1) + my_fibonacci (x - 2);
  23	FIRST_LINE + 7: }
  (gdb) next
  22	FIRST_LINE + 6:     return my_fibonacci (x - 1) + my_fibonacci (x - 2);
  (gdb) p x
  $1 = 10
  (gdb) step
  Breakpoint 1, my_fibonacci (x=9) at testsuite/jit.dg/test-fibonacci.c:19
  19	FIRST_LINE + 3:   if (x < 2)

File and line information appear to be correct, but looking at the .s
output, columns are zero, so there may be a bug in how I'm setting up
columns.

  my_fibonacci:
  .LFB0:
          .file 1 "testsuite/jit.dg/test-fibonacci.c"
          .loc 1 16 0
          .cfi_startproc
          pushq   %rbp
          .cfi_def_cfa_offset 16
          .cfi_offset 6, -16
          movq    %rsp, %rbp
          .cfi_def_cfa_register 6
          pushq   %rbx
          subq    $24, %rsp
          .cfi_offset 3, -24
          movl    %edi, -20(%rbp)
          .loc 1 19 0
          cmpl    $2, -20(%rbp)
          jge     .L2
          .loc 1 20 0
          movl    -20(%rbp), %eax
          jmp     .L4
  .L2:
          .loc 1 22 0
          movl    -20(%rbp), %eax
          subl    $1, %eax
          movl    %eax, %edi
          call    my_fibonacci@PLT

Implementation notes:

The linemap API appears to require source code locations to be created
in a very particular way, by creating references to files, and then,
within them, lines and columns, in the order the lines and columns
appear in the source file.

This is well-suited to a tokenizer, but for the JIT case we want to
allow the user to submit code without having to care about sorting
the source-code locations into order.

Hence internally we defer location creation: within the client's
code-creation hook, a gcc_jit_location represents a future location.
These are grouped by file and line, and a record is kept of associations
to be made between tree instances and gcc_jit_locations instances.
Upon return from the code-creation hook, all such locations are sorted,
and used in the linemap API to create source_location instances.
The various tree nodes then have their locations set to these instances.