MIPS R6 testsuite fixes
Modify mips.exp to skip tests if the test is running in r6 mode and the test
uses an option that is incompatible with r6 (-mdsp, -mpaired-single, etc).

Add 'isa_rev<=5' to tests that don't already check isa_rev and that
should not run on mips32r6 or mips64r6.

Add 'forbid_cpu=mips.*r6' to tests where isa_rev is already being
used (to check for <=2 for example) and cannot be used again.

When downgrading a test from mips32r6 to mips32r2 due to options in the
test that are not compatible with mips32r6, change the test to use
-mips32r2 -mnan=2008 (not just -mips32r2).

This is so that tests that include stdlib.h will not fail due to not finding
stubs-o32_hard.h.  r6 compilers only have stubs-o32_hard_2008.h, because
they are nan2008 by default.