Add gcc_jit_function_add_comment
Add gcc_jit_function_add_comment, a way to add a no-op textual comment to
the internal representation of the code.

It will be optimized away, but will be visible in some dumps
and thus may be of use when debugging how client code's internal
representation gets converted to the libgccjit IR.

Internally, it's implemented by adding a dummy label to the IR, where the
name of the label is the text of the comment, wrapped in C-style comment
delimiters.

Currently this is directly visible in the tree and gimple dumps, and
seems to survive all the way through to the end of RTL, in the form
of NOTE_INSN_DELETED_LABEL.  It doesn't make it into the final assembler
output, though.