Fix how locals are created; add BIND_EXPR
Attempting to add test coverage uncovered that locals weren't being
associated with their functions, leading to crashes later on in the
compile for various valid constructs.

To fix this, we now set the DECL_CONTEXT of locals to be the
FUNCTION_DECL.  To do so, we also need to put them within a BIND_EXPR
otherwise gimplify_var_or_parm_decl dies, with:
       !DECL_SEEN_IN_BIND_EXPR_P (decl)

Hence we now create a BIND_EXPR for every function with a body, moving
the statement list within it.

gcc/jit/
	* internal-api.h (gcc::jit::function): Add field
	"m_inner_bind_expr".
	* internal-api.c (gcc::jit::function::function): Create a BIND_EXPR
	for all non-imported functions, and put the statement list within
	it.
	(gcc::jit::function::gt_ggc_mx): Visit m_inner_bind_expr.
	(gcc::jit::function::new_local): Set the DECL_CONTEXT of the new
	local to be the function's BIND_EXPR, and prepend the new local
	to said BIND_EXPR's BIND_EXPR_VARS chain.
	(gcc::jit::function::postprocess): Set the DECL_SAVED_TREE of the
	FUNCTION_DECL to be the BIND_EXPR, rather than the statement list.
	The latter is now contained within the former.

gcc/testsuite/
	* jit.dg/test-reading-struct.c: New test, to provide test coverage
	of gcc_jit_type_get_const and gcc_jit_lvalue_access_field, in the
	process uncovering bugs in how locals were handled.
	* jit.dg/test-combination.c: Add usage of test-reading-struct.c.