Add libgccjit++.h, a C++ wrapper to the API
The libgccjit C API is relatively low-level and thus should be easy to
maintain a stable ABI for, and to wrap from many different languages.

However, it's rather verbose.

This commit adds an example of a C++ wrapper, which makes use of methods,
overloading, and inheritance, giving a much more terse API.

It's all done as small inline functions wrapped around the raw C pointers,
within one header, so it shouldn't need to be built as a separate library,
hence this shouldn't add an ABI (I hope), just an API.

As an example, I ported test-quadratic.c to it (as a new file:
test-quadratic.cc), and was able to considerably reduce the line-count
whilst fitting within a 80-column limit:

# line, word, byte counts:
$ wc \
    gcc/testsuite/jit.dg/test-quadratic.c \
    gcc/testsuite/jit.dg/test-quadratic.cc
  487  1066 12748 gcc/testsuite/jit.dg/test-quadratic.c
  369   911  9985 gcc/testsuite/jit.dg/test-quadratic.cc
  856  1977 22733 total

Caveat: I haven't yet managed to get DejaGnu to run this as part of
the test suite; I've been building/running it manually for now.

gcc/jit/
	* libgccjit++.h: New file - a C++ wrapper for the libgccjit.h API.

	* TODO.rst ("Test Suite"): New section, adding note about C++
	tests.

gcc/testsuite/
	* jit.dg/test-quadratic.cc: New file - a translation of
	test-quadratic.c to the libgccjit++.h C++ API.