Add new function in LRA: base_to_reg.
LRA emitted base+disp insn when failed to validate the address.
The validation can fail either because of an invalid offset or
invalid hard register for a mode. In mips16 case, $sp is not directly
accessible for modes lower than SImode, thus, the backend rejects it.
The LRA then inserted an add insn and used it as a new base.
This prevented using the offset for store/load for other modes
than SImode and increased the code size.

With this feature, instead of emitting an add, the base reg is
attempted to be copied into a pseudo and the old instruction
with old offset but the new base is revalidated. If it fails
it reverts the changes and lets handle the invalid address
with the original approach.