Fix global state init_p within ipa_init
Attempting to repeatedly compile with optimizations on within one
process would, under some circumstances, lead to a segfault the second
time in, here:

 #0  splay_tree_delete (sp=0x0) at ../../src/libiberty/splay-tree.c:353
 #1  0x00007ffff77f104f in propagate () at ../../src/gcc/ipa-reference.c:878
 #2  0x00007ffff77f1d23 in (anonymous namespace)::pass_ipa_reference::execute
	(this=0x72f5d0) at ../../src/gcc/ipa-reference.c:1191
 #3  0x00007ffff6fc5237 in execute_one_pass (pass=0x72f5d0) at
	../../src/gcc/passes.c:2216
 #4  0x00007ffff6fc6044 in execute_ipa_pass_list (pass=0x72f5d0) at
	../../src/gcc/passes.c:2580
 #5  0x00007ffff6c93d5f in ipa_passes () at ../../src/gcc/cgraphunit.c:2027

The crash occurred here:
 #1  0x00007ffff77f104f in propagate () at ../../src/gcc/ipa-reference.c:878
 878	    splay_tree_delete (reference_vars_to_consider);
 (gdb) p reference_vars_to_consider
 $10 = (splay_tree) 0x0

This tended to occur when working with functions intended to be inlined.

Debuggging indicated that the root-cause was the "init_p" flag within
ipa_init: it is static, presumably to allow idempotent initialization.
However, there is also cleanup, so that the 2nd time in, ipa-reference.c
is working with post-clean up state, failing to rebuild its data structures
(e.g. a NULL for "reference_vars_to_consider").

Hence this patches changes ipa initialization from being idempotent within
the lifetime of the process to being idempotent within one invocation of
toplev_main, by moving it from static in function scope to *file* scope,
adding a reset of the flag to toplev_finalize.

[That said, I *think* ipa_init is only called at most once per invocation
of toplev_main.]

Doing so allows the optimizer to inline functions in repeated invocations
of the compiler within one process.
[...provided debuginfo generation is disabled.  In this latter case, with
GCC_JIT_BOOL_OPTION_DEBUGINFO, the equivalent of -g, I'm currently running
into what I take to be a separate segfault in dwarf2out.c the second time
in.  I haven't tracked this other one down yet.]

gcc/
	* ipa-reference.c (ipa_init): Move static bool init_p from here
	to...
	(ipa_init_p): New file-scope variable, so that it can be reset
	when repeatedly invoking the compiler within one process by...
	(ipa_reference_c_finalize): New function.
	* ipa-reference.h (ipa_reference_c_finalize): New.
	* toplev.c (toplev_finalize): Invoke new function
	ipa_reference_c_finalize.