Fix state issue in gcse.c
Turning optimizations up from 0 to 3 showed a segfault on the 2nd
iteration of test-linked-list.c in get_data_from_adhoc_loc whilst
garbage-collecting.

Investigation revealed the issue to be a CFG from the previous compile
being kept alive by this GC root in gcse.c:
	  static GTY(()) rtx test_insn;

This wouldn't it itself be an issue, but one (or more) of the edges had:

  (gdb) p /x e->goto_locus
  $9 = 0x80000000

and was thus treated as an ADHOC_LOC.  Hence, this line in the edge_def's
gt_ggc_mx routine:

  8313	  tree block = LOCATION_BLOCK (e->goto_locus);

led to a call (via LOCATION_BLOCK) to get_data_from_adhoc_loc:

152	void *
153	get_data_from_adhoc_loc (struct line_maps *set, source_location loc)
154	{
155	  linemap_assert (IS_ADHOC_LOC (loc));
156	  return set->location_adhoc_data_map.data[loc & MAX_SOURCE_LOCATION].data;
157	}

but at this point, the ad-hoc location data from the previous in-process
compile no longer makes sense, and, with:
  (gdb) p set->location_adhoc_data_map
  $5 = {htab = 0x60fbb0, curr_loc = 0, allocated = 0, data = 0x0}

the read though the NULL "data" segfaults.

gcse.c appears to create test_insn on-demand as part of the implementation of
can_assign_to_reg_without_clobbers_p.

Hence it seems to make most sense to simply clear test_insn in toplev_finalize,
which this commit does.

With this, the whole test suite now runs successfully with optimizations
on, so also turn this up (in harness.h) from 0 to 3.

gcc/
	* gcse.c (gcse_c_finalize): New, to clear test_insn between
	in-process compiles.
	* gcse.h (gcse_c_finalize): New.
	* toplev.c: Include "gcse.h" so that we can...
	(toplev_finalize): Call gcse_c_finalize.

gcc/testsuite/
	* jit.dg/harness.h (set_options): Increase optimization level from
	0 to 3.