LRA: Fix the max number of generated reload insns achieved.
LRA does not change spilled pseudos to memory when calling
the secondary_reload_class hook. The current implementation
did not handle such a case and returned GR_REGS class. Then
LRA tried to satisfy the constraints but went into an infinite
reload cycling (limited to 90 reloads) as there was no alternative
that would move FP values to integer registers in CCF mode.

The problem only exists when the floating point condition code
register was spilled. The fix checks if the pseudo is in memory
in CCF mode and returns NO_REGS to indicate that no reload is needed.