Reject creating arrays of fieldless structs
Investigating a bug in a project using libgccjit revealed a problem where
fields within a struct within the JIT-created code weren't laid out in the
same places as the supposedly same struct in normally-built code:

  struct foo
  {
    int a;
    int b;
    int c;
  };

  struct bar
  {
    int d;
    struct foo e[20];
    int f; /* <<< error here */
  };

Field "f" was erroneously compiled as if "e" was only 4 bytes in size.

The root cause turned out to be that the client code was creating the
array type for
    struct foo e[20]
before the fields of "struct foo" had been set, hence the call to
layout_type for the array type would fail this conditional:
  2168  /* We need to know both bounds in order to compute the size.  */
  2169  if (index && TYPE_MAX_VALUE (index) && TYPE_MIN_VALUE (index)
  2170      && TYPE_SIZE (element))
because TYPE_SIZE(element) was 0 at that time.

Add an error for this case, to ensure that client code can only create an
array of structs once the fields of the struct have been set up.

gcc/jit/
	* internal-api.c (gcc::jit::recording::context::new_array_type):
	Reject attempts to create an array of a struct if the fields of
	the struct haven't yet been set.
	* internal-api.h (gcc::jit::recording::type::dyn_cast_struct): New
	virtual function.
	(gcc::jit::recording::struct_::dyn_cast_struct): New, overriding
	for this subclass.
	(gcc::jit::recording::struct_::get_name): New.