Allow to split the move in TFmode when generating MSA code.
*movti pattern along with the related split pattern was disabled for MSA but
the split pattern should not be disabled for TFmode. *movtf pattern instructs
the compiler to split the insn but the combiner phase then cannot find the
necessary split pattern.

Fixes an ICE found when building SPEC2006 for MIPS64 target.

To trigger it, the following code should be built with -mips64r2 -mabi=64 -mmsa:
long double a;
foo() { foo2(a);}