libgomp plugins: Remove the (OpenACC) avail function.
This idea has been to use this to see whether a class of devices is generally
available, without determining the exact number of devices that are available.
Such a distinction probably doesn't make too much sense; in the following, we'd
like to use one device anyway.

Inline the few users.

	libgomp/
	* target.h (struct ACC_dispatch_t): Remove avail_func.  Update all
	users.
	* oacc-host.c (openacc_avail): Likewise.
	* plugin-nvptx.c (PTX_avail, openacc_avail): Likewise.
	* libgomp.texi (acc_get_num_devices): Revise.

git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/gomp-4_0-branch@217141 138bc75d-0d04-0410-961f-82ee72b054a4