Change linux-common.h to make android generate PIC by default.
This code uses modified linux-android.h (not mips specific) which
is different then the ToT FSF linux-android.h.  I.e. google has a
non-mips local change in the GCC tree.  I don't know if I want this
change here because I probably don't want it included in a diff that
I apply to the Google tree.  But I do want it here to have a compilable
source tree.