Release branch only commit: Fix the sync patterns to use the ZC rather than the ZR memory constraint.
Also change the ZC memory constraint to understand the 9 bit offset
range for R6.

This fix will mean the memory constraints for the ll and sc instructions
are correctly described.  This means a memory offset can be used to
access a data item (ie. %lo (<var>)) rather than always having to load the
address into a register, and use that register within the ll/sc instruction.