Release branch only commit: Add missing comma in DRIVER_SELF_SPECS
The mti-linux.h definition of DRIVER_SELF_SPECS was missing a comma
which meant the last entry of BASE_DRIVER_SELF_SPECS was merged
with the next entry from LINUX_DRIVER_SELF_SPECS. This is mostly not
a problem but commas throughout is cleaner. Whitespace will separate
the options currently.