Remove the prototypes for builtins. Force the type for {,U}V2DI mode to {,unsigned} long long.
Changing the type is important for the removed builtins as when using and
compiling for -mips64* with -mabi=64 we ran into incompatibility between
the types. On 64-bit architecture, builtins are initialized to expect
'long int' instead of 'long long' for a vector with 64-bit elements, hence,
an error.  Changing the expected type to 'long long' appears to be sensible
since I'd expect a guaranteed 64-bit size and the same type regardless of
the ABI used.