vec::pop should return an object instead of a reference
vec::pop () used to return a reference to an object that after the call
is out side of the range 0 - length -1.  So conseptually this was kind
of a use after free since otherwise a vec implementation would be
allowed to shrink the internal storage to not include the memory
referenced by the return value of pop.  This also means if we make vec
call constructors and destructors then the object whose reference will
be returned will be destructed before the caller gets the reference.
The stl avoids this problem by having pop_back () return void, however
there are several sites where vec's API is simpler to use.  This leaves
us with making a copy of the object about to be removed from the vec and
returning that.  This can be rather painful if objects in the vec are
slow to copy, however it probably isn't a big deal practice for a couple
reasons.  First the vast majority of vecs are storing word sized objects
that are POD.  Next in looking at the callers of pop () one sees few if
any sites that assign the return value to a T &, instead they all assign
to a T, meaning a copy happens already.

gcc/

2015-03-22  Trevor Saunders  <tbsaunde+gcc@tbsaunde.org>

	* vec.h (vec::pop): Change return type from T & to T.