Don't increase the cost of move insn if curr_insn got equiv substitution.
In the remaining cases where the integers are placed into FP registers, LRA
was rematerializing registers which replacement by equiv subsitution was
detected by LRA as a move insn and penalized to avoid infinite cycling.
Because of the increase of the cost, the GP-to-GP alternative was rejected and
FP-to-GP alternative was chosen instead.

The patch attempts to make an exception for instructions which got
the equivalence substitution in the process. It appears that not only the issue
with integers in FP regs disappears but also some (unneccesary) moves are not
inserted and it's more visible with bare-metal toolchain and mips64r6.